Usando long para numeros maiores

+++ calculator.c
@@ -2,14 +2,14 @@
 #include <stdio.h>
 
 int main(void){
-    int dollars = 1;
+    long dollars = 1;
     while(true){
-        char c = get_char("Here's $%i. Double it and give it to the next person? ", dollars);
+        char c = get_char("Here's $%li. Double it and give it to the next person? ", dollars);
         if (c == 'y' || c == 'Y'){
             dollars *= 2;
         }else{
             break;
         }
     }
-    printf("Here's $%i", dollars);
+    printf("Here's $%li", dollars);
 }
